Add navigation widget (WIP)
Added a navigation compass to assist the reader with L/R A/P I/S. This works by reading the acquisition angles from the parameter file and, for now, assumes supine head first. It also does not update when the image rotate/flip buttons are used so be cautious.

Added unwrapping to the phase image generation.

Moved the noise estimation for multichannel reconstruction into the combine channel function as it was throwing errors when incompatible data sets were loaded (typically unfinished single channel scans).